fix: AuthDto, MemberDto 변환
